update formatting of attributes (price, volume, quantity) after submitting an item

+++ db/migrate/20141023094955_create_items.rb
@@ -2,7 +2,7 @@ class CreateItems < ActiveRecord::Migration
   def change
     create_table :items do |t|
       t.string :category
-      t.integer :volume
+      t.float :volume
       t.string :volume_type
       t.float :price
       t.integer :quantity
